\name{Austria}
\alias{Austria}
\title{Emissions in Austria in 10 air pollutants divided by the ratio between the gross value added of manufacturing and the gross value-added sourced from all economic activities (years 2008-2015)}

\description{The dataset is provided by the AEA accounts (Air Emissions Account) collected yearly by EUROSTAT for the seven-year period 2008-2015. AEA accounts report emissions assigned to the country according to the residence principle (i.e. the residence of the operator causing the emission). The list of pollutants included in the analysis embraces 10 air pollutants that are particularly harmful for human health and environmental balance when highly concentrated in the atmosphere (see Caravaggio et al. 2019)}

\usage{data(Austria)}

\references{\url{https://doi.org/10.1016/j.jclepro.2019.02.115}}

\details{Caravaggio, N., Caravella, S., Ishizaka, A., & Resce, G. (2019). Beyond CO2: a multi-criteria analysis of air pollution in Europe. Journal of Cleaner Production.}

\author{Nicola Caravaggio, Serenella Caravella, Alessio Ishizaka, Giuliano Resce, Francesco Vidoli}
\examples{data(Austria)}