% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize.R
\name{normalize}
\alias{normalize}
\title{Normalization}
\usage{
normalize(data, beneficial_col = NULL)
}
\arguments{
\item{data}{The data that needs to be normalized}

\item{beneficial_vars}{The variables whose highest value is considered
to be maximum and lowest value is minimum.}
}
\value{
A dataframe with normalized values
}
\description{
This function normalizes the input data. It uses min max normalization.
}
