% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preference_function.R
\name{pref_func}
\alias{pref_func}
\title{Preference Function}
\usage{
pref_func(row_wise_difference, preference_function = "simple", p = 1,
  q = 0, s = 0.5)
}
\value{
An output data from preference function.
}
\description{
This function applies preference function to the pair wise
difference data.
s
}
