% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row_wise_difference.R
\name{row_wise_difference}
\alias{row_wise_difference}
\title{Pairwise differences}
\usage{
row_wise_difference(normalized_data)
}
\arguments{
\item{data}{The data that needs to be normalized}

\item{beneficial_vars}{The variables whose highest value is considered
to be maximum and lowest value is minimum.}
}
\value{
A dataframe with differences between all pairs of alternatives
across all criteria.
}
\description{
This function calculates the differences for
each pair of alternatives using permutation.
}
